requirements.txt
argparse
futures
insightface
jupyterlab
keras
librosa
matplotlib
mxnet
mxnet-cu100
numpy
opencv-python
pandas
scipy
setuptools==41.0.0
six==1.12.0
tensorflow-gpu==2.1.0
tqdm
